refactor: declare getInstance() for MenuSite
