+++ src/main/java/com/holub/ui/MenuSite.java
@@ -1,5 +1,6 @@
 package com.holub.ui;
 
+import com.holub.system.Clock;
 import java.io.*;
 import java.util.*;
 import java.util.logging.*;
@@ -99,8 +100,25 @@ import javax.swing.*;
 
 public final class MenuSite {
 
-  private static JFrame menuFrame = null;
-  private static JMenuBar menuBar = null;
+  private JFrame menuFrame = null;
+  public JFrame getMenuFrame() {
+    return menuFrame;
+  }
+  private JMenuBar menuBar = null;
+  public void SetMenuBar(JMenuBar menuBar) {
+    this.menuBar = menuBar;
+  }
+  public JMenuBar getMenuBar() {
+    return menuBar;
+  }
+  private static MenuSite instance;
+
+  public synchronized static MenuSite getInstance() {
+    if (instance == null) {
+      instance = new MenuSite();
+    }
+    return instance;
+  }
 
   /*** The "requesters" table keeps track of who requested which
    * menu items. It is indexed by requester and contains a
@@ -158,8 +176,7 @@ public final class MenuSite {
    *	will appear on the screen (left to right).
    */
 
-  private static final LinkedList menuBarContents =
-      new LinkedList();
+  public final LinkedList menuBarContents = new LinkedList();
 
   /*** ***************************************************************
    * MenuSite is a singleton. A private constructor prevents
@@ -175,7 +192,7 @@ public final class MenuSite {
    *  @throws AssertionException if the menu hasn't been established.
    */
 
-  private static boolean valid() {
+  private boolean valid() {
     assert menuFrame != null : "MenuSite not established";
     assert menuBar != null : "MenuSite not established";
     return true;
@@ -187,7 +204,7 @@ public final class MenuSite {
    * (Most of these will throw a {@link NullPointerException}
    * if you try to use them when no menu site has been established.)
    */
-  public synchronized static void establish(JFrame container) {
+  public synchronized void establish(JFrame container) {
     assert container != null;
     assert menuFrame == null :
         "Tried to establish more than one MenuSite";
@@ -254,7 +271,7 @@ public final class MenuSite {
    *  		(e.g. has spaces in it) or if the specifier identifies
    *  		an existing line item (as compared to a menu).
    */
-  public static void addMenu(Object requester, String menuSpecifier) {
+  public void addMenu(Object requester, String menuSpecifier) {
     createSubmenuByName(requester, menuSpecifier);
   }
 
@@ -304,7 +321,7 @@ public final class MenuSite {
    *  @see #addMenu
    *  @see #mapNames
    */
-  public static void addLine(Object requester,
+  public void addLine(Object requester,
       String toThisMenu,
       String name,
       ActionListener listener) {
@@ -361,7 +378,7 @@ public final class MenuSite {
    *  the menu itself using {@link #addMenu addMenu(...)}.
    */
 
-  public static void removeMyMenus(Object requester) {
+  public void removeMyMenus(Object requester) {
     assert requester != null;
     assert valid();
 
@@ -387,7 +404,7 @@ public final class MenuSite {
    * @param enable true to enable all the requester's menu items.
    *
    */
-  public static void setEnable(Object requester, boolean enable) {
+  public void setEnable(Object requester, boolean enable) {
     assert requester != null;
     assert valid();
 
@@ -451,7 +468,7 @@ public final class MenuSite {
    * 						doesn't exist.
    */
 
-  public static JMenuItem getMyMenuItem(Object requester,
+  public JMenuItem getMyMenuItem(Object requester,
       String menuSpecifier, String name) {
     assert requester != null;
     assert menuSpecifier != null;
@@ -490,7 +507,7 @@ public final class MenuSite {
    * doesn't exist, create it.
    * @see #addMenu
    */
-  private static JMenu createSubmenuByName(Object requester,
+  private JMenu createSubmenuByName(Object requester,
       String menuSpecifier) {
     assert requester != null;
     assert menuSpecifier != null;
@@ -737,7 +754,7 @@ public final class MenuSite {
    * if there are no menus associated with the requester at
    * present.
    */
-  private static Collection menusAddedBy(Object requester) {
+  private Collection menusAddedBy(Object requester) {
     assert requester != null : "Bad argument";
     assert requesters != null : "No requesters";
     assert valid();
@@ -750,175 +767,6 @@ public final class MenuSite {
     return menus;
   }
 
-  /*** ***********************************************************
-   * An Item makes the association between a line item or
-   * submenu and the MenuBar or Menu that contains it. You can
-   * ask an Item to add or remove itself from its container.
-   * All the weirdness associated with help menus is handled
-   * here.
-   */
-  private static final class Item {
-
-    // private JMenuItem  item;
-    private Component item;
-
-    private String parentSpecification; // of JMenu or of
-    // JMenuItem's parent
-    private MenuElement parent;           // JMenu or JMenuBar
-    private boolean isHelpMenu;
-
-    public String toString() {
-      StringBuffer b = new StringBuffer(parentSpecification);
-      if (item instanceof JMenuItem) {
-        JMenuItem i = (JMenuItem) item;
-        b.append(":");
-        b.append(i.getName());
-        b.append(" (");
-        b.append(i.getText());
-        b.append(")");
-      }
-      return b.toString();
-    }
-
-    /*------------------------------------------------------------*/
-
-    private boolean valid() {
-      assert item != null : "item is null";
-      assert parent != null : "parent is null";
-      return true;
-    }
-
-    /*** Create a new Item. If the JMenuItem's name is the
-     *  string "help" then it's assumed to be the help menu and
-     *  is treated specially. Note that several help menus can
-     *  be added to a site: They'll be stacked up at the far
-     *  right in the reverse order of addition. Similarly
-     *  file menus are stacked up at the far left.
-     *
-     *  @param item     the item being added
-     *  @param parent   The menu bar or a menu that
-     *  				 contains the current item. Must
-     *  				 be a JMenuBar or a JMenu.
-     */
-
-    public Item(Component item, MenuElement parent,
-        String parentSpecification) {
-      assert parent != null;
-      assert parent instanceof JMenu || parent instanceof JMenuBar
-          : "Parent must be JMenu or JMenuBar";
-
-      this.item = item;
-      this.parent = parent;
-      this.parentSpecification = parentSpecification;
-      this.isHelpMenu =
-          (item instanceof JMenuItem)
-              && (item.getName().compareToIgnoreCase("help") == 0);
-
-      assert valid();
-    }
-
-    public boolean specifiedBy(String specifier) {
-      return parentSpecification.equals(specifier);
-    }
-
-    public Component item() {
-      return item;
-    }
-
-    /*** ******************************************************
-     * Attach a menu item to it's parent (either a menu
-     * bar or a menu). Items are added at the end of the
-     * <code>menuBarContents</code> list unless a help
-     * menu exists, in which case items are added at
-     * the penultimate position.
-     */
-
-    public final void attachYourselfToYourParent() {
-      assert valid();
-
-      if (parent instanceof JMenu) {
-        ((JMenu) parent).add(item);
-      } else if (menuBarContents.size() <= 0) {
-        menuBarContents.add(this);
-        ((JMenuBar) parent).add(item);
-      } else {
-        Item last = (Item) (menuBarContents.getLast());
-        if (!last.isHelpMenu) {
-          menuBarContents.addLast(this);
-          ((JMenuBar) parent).add(item);
-        } else  // remove the help menu, add the new
-        {    // item, then put the help menu back
-          // (following the new item).
-
-          menuBarContents.removeLast();
-          menuBarContents.add(this);
-          menuBarContents.add(last);
-
-          if (parent == menuBar) {
-            parent = regenerateMenuBar();
-          }
-        }
-      }
-    }
-
-    /*** ******************************************************
-     * Remove the current menu item from its parent
-     * (either a menu bar or a menu). The Item is invalid
-     * after it's detached, and should be discarded.
-     */
-    public void detachYourselfFromYourParent() {
-      assert valid();
-
-      if (parent instanceof JMenu) {
-        ((JMenu) parent).remove(item);
-      } else // the parent's the menu bar.
-      {
-        menuBar.remove(item);
-        menuBarContents.remove(this);
-        regenerateMenuBar(); // without me on it
-
-        parent = null;
-      }
-    }
-
-    /*** ******************************************************
-     * Set or reset the "disabled" state of a menu item.
-     */
-
-    public void setEnableAttribute(boolean on) {
-      if (item instanceof JMenuItem) {
-        JMenuItem item = (JMenuItem) this.item;
-        item.setEnabled(on);
-      }
-    }
-
-    /*** ******************************************************
-     * Replace the old menu bar with a new one that reflects
-     * the current state of the <code>menuBarContents</code>
-     * list.
-     */
-    private JMenuBar regenerateMenuBar() {
-      assert valid();
-
-      // Create the new menu bar and populate it from
-      // the current-contents list.
-
-      menuBar = new JMenuBar();
-      ListIterator i = menuBarContents.listIterator(0);
-      while (i.hasNext()) {
-        menuBar.add(((Item) (i.next())).item);
-      }
-
-      // Replace the old menu bar with the new one.
-      // Calling setVisible causes the menu bar to be
-      // redrawn with a minimum amount of flicker. Without
-      // it, the redraw doesn't happen at all.
-
-      menuFrame.setJMenuBar(menuBar);
-      menuFrame.setVisible(true);
-      return menuBar;
-    }
-  }
 
   /*** ***************************************************************
    * This class holds methods of interest only when you're
